[unity] v2 - combine ModuleManager, PromiseHandler, Inspector to BackendEnv
